+++ src/UnitySampleProject/Assets/Scripts/Prefabs/Minter/Mint1155.cs
@@ -6,6 +6,7 @@ using System;
 using Web3Unity.Scripts.Library.Ethers.Transactions;
 using Nethereum.Hex.HexTypes;
 using Newtonsoft.Json;
+using System.Numerics;
 
 public class Mint1155 : MonoBehaviour
 {
@@ -30,12 +31,12 @@ public class Mint1155 : MonoBehaviour
         RedeemVoucherTxModel.Response voucherResponse = await EVM.CreateRedeemTransaction(Web3Accessor.Web3, chainConfig.Chain, chainConfig.Network, voucherArgs, "1155", nftAddress, voucherResponse1155.receiver);
         var txRequest = new TransactionRequest
         {
-            ChainId = new HexBigInteger(int.Parse(chainConfig.ChainId)),
+            ChainId = new HexBigInteger(BigInteger.Parse(chainConfig.ChainId)),
             To = voucherResponse.tx.to,
             Value = new HexBigInteger(voucherResponse.tx.value),
             Data = voucherResponse.tx.data,
-            GasLimit = new HexBigInteger(int.Parse(voucherResponse.tx.gasLimit)),
-            GasPrice = new HexBigInteger(int.Parse(voucherResponse.tx.gasPrice)),
+            GasLimit = new HexBigInteger(BigInteger.Parse(voucherResponse.tx.gasLimit)),
+            GasPrice = new HexBigInteger(BigInteger.Parse(voucherResponse.tx.gasPrice)),
         };
         var response = await Web3Accessor.Web3.TransactionExecutor.SendTransaction(txRequest);
         Debug.Log(JsonConvert.SerializeObject(response));
